requirements.txt
langchain==0.2.6
langchain-core==0.2.10
langchain-google-genai==1.0.6
langchain-community==0.2.6
langsmith==0.1.81
chromadb==0.5.3
pandas
pypdf
streamlit
python-dotenv